Fix bin/sp-ization-benchmarking build
