+++ bin/sp-ization-benchmarking/CMakeLists.txt
@@ -1,6 +1,8 @@
 ff_add_executable(
   NAME
     sp-ization-benchmarking
+  PRIVATE_INCLUDE
+    include/
   SRC_PATTERNS
     *.cc
   DEPS
