+++ bin/sp-ization-benchmarking/include/sp-ization-benchmarking/distributions.h
@@ -1,5 +1,5 @@
-#ifndef _FLEXFLOW_DISTRIBUTIONS_H
-#define _FLEXFLOW_DISTRIBUTIONS_H
+#ifndef _FLEXFLOW_BIN_SP_IZATION_BENCHMARKING_INCLUDE_SP_IZATION_BENCHMARKING_DISTRIBUTIONS_H
+#define _FLEXFLOW_BIN_SP_IZATION_BENCHMARKING_INCLUDE_SP_IZATION_BENCHMARKING_DISTRIBUTIONS_H
 
 #include "utils/graph/node/node.dtg.h"
 #include <random>
