+++ bin/sp-ization-benchmarking/src/sp-ization-benchmarking/distributions.cc
@@ -1,4 +1,4 @@
-#include "distributions.h"
+#include "sp-ization-benchmarking/distributions.h"
 
 namespace FlexFlow {
